refactor: move undo/ into engine/ (cross-cutting infrastructure)
Undo (snapshot, rollback) is a low-level primitive used by refactor,
auto-fix, and the undo CLI command. It's cross-cutting infrastructure
that belongs in engine alongside shell, command, temp, pipeline —
not as a peer of domain capabilities like deploy and release.

The CLI layer doesn't need to mirror core structure 1:1.
commands/undo.rs still exists as a thin wrapper around the core API.

Import paths updated via homeboy refactor rename.

765 tests pass, cargo fmt clean, zero new warnings.
