+++ src/commands/undo.rs
@@ -1,7 +1,7 @@
 use clap::{Args, Subcommand};
 use serde::Serialize;
 
-use homeboy::undo;
+use homeboy::engine::undo;
 
 use super::CmdResult;
