+++ src/core/refactor/auto/apply.rs
@@ -3,7 +3,7 @@ use crate::core::refactor::decompose;
 use crate::core::refactor::plan::generate::primary_type_name_from_declaration;
 use crate::core::refactor::plan::verify::rewrite_callers_after_dedup;
 use crate::core::refactor::shared::detect_language;
-use crate::core::undo::InMemoryRollback;
+use crate::engine::undo::InMemoryRollback;
 use crate::refactor::auto::{
     ApplyChunkResult, ApplyOptions, ChunkStatus, DecomposeFixPlan, Fix, FixResult, Insertion,
     InsertionKind, NewFile,
